Experimental swap chain drawing
This commit moves from Hwnd render targets to DXGI swap chains.
Some things are hacky, but overall it mostly works.

Some things don't work well though. Rapid resizing of the window
can cause out of memory errors. Resizing is not smooth.

This commit also moves to the 0.3 branch of winapi, as not all
of the dxgi methods were available on 0.2 during experimentation.